osd/ReplicatedPG: For CACHEMODE_READONLY, don't allow update existing object in cache pool.
As doc descript:"For CACHMODE_READONLY,we copy data to the cache pool(s) on read. Writes are
forwarded to the original data pool.Stale data is expired from the cache
pools based on the defined policy."
But now we can update(rewrite) existing object in cache pool.
So fix it, for write don't care existing in cache pool only redirect to
base pool.

Signed-off-by: Jianpeng Ma <jianpeng.ma@intel.com>