rgw: completion of multipart upload leaves delete marker
The multipart upload meta object is deleted when the multipart upload
is completed. When the bucket is versioned, it needs to be deleted
from the bucket index rather than go through versioning delete process
that adds a delete marker to the bucket index rather than simply
removing it from the bucket index.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>